AOP 적용을 위한 EventUpsert 분리(#210)
